Split sql store functions into file per entity.
